requirements.txt
blessed==1.19.1
inquirer==2.10.0
python-dotenv==0.20.0
python-editor==1.0.4
readchar==3.1.0
six==1.16.0
wcwidth==0.2.5
